src/requirements.txt
psutil==7.1.3
discord.py==2.6.4